fix: some small fixups
